# microapp-customers/src/main/resources/openapi/customer-api.yml
openapi: 3.0.1
info:
  title: Customer API
  description: This is an API to handle requests regarding customers
  version: 1.0.0
paths:
  /customers:
    get:
      summary: Retrieve all customers
      operationId: getCustomers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      summary: Add customer
      operationId: createCustomer
      requestBody:
        description: Create customer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
    put:
      summary: Update customer
      operationId: updateCustomer
      requestBody:
        description: Update customer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /customers/{id}:
    get:
      summary: Retrieve customer by id
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    delete:
      summary: Delete customer by id
      operationId: deleteCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    Customer:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string